project/pyproject.toml
[tool.poetry]
name = "LanggraphAgents"
version = "0.1.0"
description = "Python Langgraph coding and research agents"
authors = ["iangalvao <iansgalvao@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.3.25"
langgraph = "^0.4.8"
pandas = "^2.3.0"
python-dotenv = "^1.1.0"
langchain-community = "^0.3.25"
openai = "^1.86.0"
langchain-openai = "^0.3.23"

[tool.poetry.dev-dependencies]
mlflow = "^3.1.0"
jupyter = "^1.1.1"
ipykernel = "^6.29.5"
pytest = "^8.4.0"
mypy = "^1.16.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.11"
mypy_path = ["./"]
